server-flask/requirements.txt
Flask==3.1.2
flask-cors==4.0.0
PyMuPDF==1.26.7
cloudinary==1.44.1
pdfplumber==0.11.9
python-dotenv==1.2.1